+++ src/RcppExports.cpp
@@ -1160,6 +1160,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// get_batch_size_
+int get_batch_size_(SEXP sim_);
+RcppExport SEXP _simmer_get_batch_size_(SEXP sim_SEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
+    rcpp_result_gen = Rcpp::wrap(get_batch_size_(sim_));
+    return rcpp_result_gen;
+END_RCPP
+}
 // get_capacity_
 SEXP get_capacity_(SEXP sim_, const std::vector<std::string>& names);
 RcppExport SEXP _simmer_get_capacity_(SEXP sim_SEXP, SEXP namesSEXP) {
@@ -1560,6 +1571,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_simmer_get_name_", (DL_FUNC) &_simmer_get_name_, 1},
     {"_simmer_get_attribute_", (DL_FUNC) &_simmer_get_attribute_, 3},
     {"_simmer_get_prioritization_", (DL_FUNC) &_simmer_get_prioritization_, 1},
+    {"_simmer_get_batch_size_", (DL_FUNC) &_simmer_get_batch_size_, 1},
     {"_simmer_get_capacity_", (DL_FUNC) &_simmer_get_capacity_, 2},
     {"_simmer_get_capacity_selected_", (DL_FUNC) &_simmer_get_capacity_selected_, 2},
     {"_simmer_get_queue_size_", (DL_FUNC) &_simmer_get_queue_size_, 2},