+++ NEWS.md
@@ -1,4 +1,4 @@
-# simmer 4.4.4.4
+# simmer 4.4.4.5
 
 ## Minor changes and fixes
 
@@ -6,10 +6,12 @@
   function environments for generators too (#279).
 - Fix file removal on object destruction for file-backed monitors (#277).
 - Remove cap on the number of nested branches for trajectory printing (#262).
+- Fix duplicate time-based batch trigger (as part of #263).
 
 ## New features
 
 - Add arithmetic support for `schedule` objects (#272).
+- New `get_batch_size()` getter allows a batch to retrieve its size (#263).
 
 # simmer 4.4.4